[fix][broker] Fix issue that topic policies was deleted after a sub topic deleted, even if the partitioned topic still exists (#24350)
(cherry picked from commit 8db6a3b9b0d03bbaf10d825c82f3fa8f9bd6a5cb)
(cherry picked from commit 0650d927f7a9e248693ffa4de08f1bb6ae99268f)

+++ pulsar-broker/src/main/java/org/apache/pulsar/broker/service/AbstractTopic.java
@@ -1298,6 +1298,9 @@ public abstract class AbstractTopic implements Topic, TopicPolicyListener<TopicP
     }
 
     public CompletableFuture<Void> deleteTopicPolicies() {
+        if (TopicName.get(getName()).isPartitioned()) {
+            return CompletableFuture.completedFuture(null);
+        }
         return brokerService.deleteTopicPolicies(TopicName.get(topic));
     }
